+++ oshdb-api/src/test/java/org/heigit/ohsome/oshdb/api/tests/TestHelpersOSMEntitySnapshotView.java
@@ -89,7 +89,7 @@ public class TestHelpersOSMEntitySnapshotView {
   @Test
   public void testCount() throws Exception {
     // single timestamp
-    SortedMap<OSHDBTimestamp, Integer> result1 = this.createMapReducer()
+    SortedMap<OSHDBTimestamp, Long> result1 = this.createMapReducer()
         .timestamps(timestamps1)
         .osmType(OSMType.WAY)
         .osmTag("building", "yes")
@@ -100,7 +100,7 @@ public class TestHelpersOSMEntitySnapshotView {
     assertEquals(42, result1.get(result1.firstKey()).intValue());
 
     // many timestamps
-    SortedMap<OSHDBTimestamp, Integer> result2 = this.createMapReducer()
+    SortedMap<OSHDBTimestamp, Long> result2 = this.createMapReducer()
         .timestamps(timestamps72)
         .osmType(OSMType.WAY)
         .osmTag("building", "yes")
@@ -112,7 +112,7 @@ public class TestHelpersOSMEntitySnapshotView {
     assertEquals(42, result2.get(result2.lastKey()).intValue());
 
     // total
-    Integer result3 = this.createMapReducer()
+    Long result3 = this.createMapReducer()
         .timestamps(timestamps1)
         .osmType(OSMType.WAY)
         .osmTag("building", "yes")
@@ -121,7 +121,7 @@ public class TestHelpersOSMEntitySnapshotView {
     assertEquals(42, result3.intValue());
 
     // custom aggregation identifier
-    SortedMap<Boolean, Integer> result4 = this.createMapReducer()
+    SortedMap<Boolean, Long> result4 = this.createMapReducer()
         .timestamps(timestamps1)
         .osmType(OSMType.WAY)
         .osmTag("building", "yes")
