+++ oshdb-api/src/test/java/org/heigit/ohsome/oshdb/api/tests/TestMapAggregateByGeometry.java
@@ -128,7 +128,7 @@ public class TestMapAggregateByGeometry {
 
   @Test
   public void testZerofill() throws Exception {
-    SortedMap<String, Integer> resultZerofilled = createMapReducerOSMEntitySnapshot()
+    SortedMap<String, Long> resultZerofilled = createMapReducerOSMEntitySnapshot()
         .timestamps(timestamps1)
         .aggregateByGeometry(getSubRegions())
         .count();
