+++ oshdb-api/src/test/java/org/heigit/ohsome/oshdb/api/tests/TestMapAggregateByIndex.java
@@ -109,7 +109,7 @@ public class TestMapAggregateByIndex {
   @Test
   public void testZerofill() throws Exception {
     // partially empty result
-    SortedMap<Long, Integer> result = createMapReducerOSMContribution()
+    SortedMap<Long, Long> result = createMapReducerOSMContribution()
         .timestamps(timestamps72)
         .osmEntityFilter(entity -> entity.getId() == 617308093)
         .aggregateBy(
@@ -120,8 +120,8 @@ public class TestMapAggregateByIndex {
 
     assertEquals(2, result.entrySet().size());
     assertEquals(true, result.containsKey(-1L));
-    assertEquals(0, (int) result.get(-1L));
-    assertEquals(7, (int) result.get(617308093L));
+    assertEquals(0, result.get(-1L).intValue());
+    assertEquals(7, result.get(617308093L).intValue());
 
     // totally empty result
     result = createMapReducerOSMContribution()
@@ -135,12 +135,12 @@ public class TestMapAggregateByIndex {
 
     assertEquals(1, result.entrySet().size());
     assertEquals(true, result.containsKey(-1L));
-    assertEquals(0, (int) result.get(-1L));
+    assertEquals(0, result.get(-1L).intValue());
   }
 
   @Test
   public void testMultiple2() throws Exception {
-    SortedMap<OSHDBCombinedIndex<Long, OSMType>, Integer> result =
+    SortedMap<OSHDBCombinedIndex<Long, OSMType>, Long> result =
         createMapReducerOSMEntitySnapshot()
             .timestamps(timestamps1)
             .osmEntityFilter(entity -> entity.getId() == 617308093)
@@ -149,14 +149,14 @@ public class TestMapAggregateByIndex {
             .count();
 
     assertEquals(1, result.entrySet().size());
-    assertEquals(1, (int) result.get(new OSHDBCombinedIndex<>(617308093L, OSMType.NODE)));
-    SortedMap<Long, SortedMap<OSMType, Integer>> nestedResult = OSHDBCombinedIndex.nest(result);
-    assertEquals(1, (int) nestedResult.get(617308093L).get(OSMType.NODE));
+    assertEquals(1, result.get(new OSHDBCombinedIndex<>(617308093L, OSMType.NODE)).intValue());
+    SortedMap<Long, SortedMap<OSMType, Long>> nestedResult = OSHDBCombinedIndex.nest(result);
+    assertEquals(1, nestedResult.get(617308093L).get(OSMType.NODE).intValue());
   }
 
   @Test
   public void testMultiple3() throws Exception {
-    SortedMap<OSHDBCombinedIndex<OSHDBCombinedIndex<Long, OSMType>, Integer>, Integer> result =
+    SortedMap<OSHDBCombinedIndex<OSHDBCombinedIndex<Long, OSMType>, Integer>, Long> result =
         createMapReducerOSMEntitySnapshot()
             .timestamps(timestamps1)
             .osmEntityFilter(entity -> entity.getId() == 617308093)
@@ -166,8 +166,8 @@ public class TestMapAggregateByIndex {
             .count();
 
     assertEquals(1, result.entrySet().size());
-    SortedMap<Long, SortedMap<OSMType, SortedMap<Integer, Integer>>> nestedResult =
+    SortedMap<Long, SortedMap<OSMType, SortedMap<Integer, Long>>> nestedResult =
         OSHDBCombinedIndex.nest(OSHDBCombinedIndex.nest(result));
-    assertEquals(1, (int) nestedResult.get(617308093L).get(OSMType.NODE).get(165061));
+    assertEquals(1, nestedResult.get(617308093L).get(OSMType.NODE).get(165061).intValue());
   }
 }
